feat(cold-archive.onboarding): pr fixes
ref: MANAGER-9181
