+++ packages/manager/modules/pci/src/projects/project/storages/cold-archive/archives/onboarding/onboarding.constants.js
@@ -1,22 +1,6 @@
-export const GUIDES = [
-  {
-    id: 'cold-archive-getting-started',
-    link:
-      'https://docs.ovh.com/fr/storage/s3/getting-started-with-cold-archive/',
-  },
-  {
-    id: 's3-getting-started',
-    link: 'https://docs.ovh.com/gb/en/storage/s3/getting-started-with-s3/',
-  },
-  {
-    id: 's3-limitaion',
-    link: 'https://docs.ovh.com/fr/storage/s3/limitations/',
-  },
-];
-
 export const TRACKING = {
   ONBOARDING_PAGE_VISITE: 'onboarding',
   ONBOARDING_CREATE_USER: 'add_a_user',
 };
 
-export default { GUIDES, TRACKING };
+export default { TRACKING };
