+++ packages/manager/modules/pci/src/projects/project/storages/cold-archive/archives/onboarding/onboarding.html
@@ -4,8 +4,8 @@
 ></pci-announcement-banner>
 
 <manager-on-boarding-layout
-    guides="$ctrl.guides"
-    image-source="$ctrl.illustration"
+    data-guides="$ctrl.guides"
+    data-image-source="$ctrl.illustration"
     data-on-guide-click="$ctrl.onDocumentationClick(guide)"
 >
     <h1
