+++ packages/manager/modules/pci/src/projects/project/storages/cold-archive/cold-archive.service.js
@@ -1,4 +1,3 @@
-/* eslint-disable no-console */
 import 'moment';
 
 export default class PciStoragesColdArchiveService {
