+++ packages/manager/modules/pci/src/projects/project/storages/cold-archive/cold-archives.constants.js
@@ -1,4 +1,20 @@
 export const COLD_ARCHIVE_TRACKING_PREFIX =
   'PublicCloud::pci::projects::project::storages::cold_archive';
 
-export { COLD_ARCHIVE_TRACKING_PREFIX as default };
+export const GUIDES = [
+  {
+    id: 'cold-archive-getting-started',
+    link:
+      'https://docs.ovh.com/fr/storage/s3/getting-started-with-cold-archive/',
+  },
+  {
+    id: 's3-getting-started',
+    link: 'https://docs.ovh.com/gb/en/storage/s3/getting-started-with-s3/',
+  },
+  {
+    id: 's3-limitaion',
+    link: 'https://docs.ovh.com/fr/storage/s3/limitations/',
+  },
+];
+
+export default { COLD_ARCHIVE_TRACKING_PREFIX, GUIDES };
